How to commit long messages, choose editors, and edit last commit message.
Use 'git commit' to open an editor and write a commit with title
and body. The title will be the first line and the body will be all
the rest after the blank line folowwing the subject (title).
To set your prefered editor, instead of vi, run the command
'export EDITOR=<editor-command-line-call>'. For instance,
if you use sublime and it's defined to run by just typing
'subl' on the command line, the command would be
'export EDITOR=subl'.

Correction: Maybe a more standard way is 'git config --global core.editor "subl -n -w"'.
The -n opens a new window, the -w tells sublime to wait for the file to be closed
before returning.

Also found out that 'git commit --ammend' lets you
edit the latest commit message.