Creating branches and pushing them to remote.
To create a new branch use 'git branch <branch-name>'.
To jump to this new branch, do 'git checkout <branch-name>'.
To create a new branch and go to it in one shot, do 'git checkout -b <branch-name>'.

To push a branch created locally that doesn't exist in remote yet,
do 'git push origin <branch-name>'.

There are also different strategies to push to remote.
You can push all branchs to remote branchs with the same name
using the strategy 'matching'. Use 'git config --global push.default matching'.

You can use the strategy 'simple', to push only the given branch and only
if there is a branch in the remote with the same name.
'git push origin <branch-name>'.

You can also push only the current branch. It will be accepted
as long as there is a branch with the same name in the remote.
'git push'.